+++ src/main/java/at/fhv/master/laendleenergy/view/UserController.java
@@ -12,6 +12,9 @@ import jakarta.ws.rs.core.Context;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
 import jakarta.ws.rs.core.SecurityContext;
+import org.eclipse.microprofile.jwt.JsonWebToken;
+
+
 import java.security.Principal;
 
 @Path("/user")
@@ -20,6 +23,8 @@ public class UserController {
 
     @Inject
     UserService userService;
+    @Inject
+    JsonWebToken jwt;
 
     @POST
     @Consumes(MediaType.APPLICATION_JSON)
@@ -36,42 +41,61 @@ public class UserController {
     }
 
     @DELETE
-    @Path("/delete/{userId}")
+    @Path("/delete")
     @Produces(MediaType.APPLICATION_JSON)
     @Authenticated
-    public Response deleteUser(String userId) {
-        try {
-            userService.deleteUser(userId);
-            return Response.ok(true).build();
-        } catch (UserNotFoundException e) {
-            return Response.status(Response.Status.NOT_FOUND).build();
+    public Response deleteUser() {
+        boolean hasJWT = jwt.getClaimNames() != null;
+
+        if (hasJWT && jwt.containsClaim("memberId")) {
+            String userId = jwt.getClaim("memberId");
+            try {
+                userService.deleteUser(userId);
+                return Response.ok(true).build();
+            } catch (UserNotFoundException e) {
+                return Response.status(Response.Status.NOT_FOUND).build();
+            }
         }
+        return Response.status(Response.Status.UNAUTHORIZED).build();
     }
 
     @GET
-    @Path("/get/all")
+    @Path("/getAll")
     @Produces(MediaType.APPLICATION_JSON)
     @Authenticated
     public Response getAllUsers() {
-        try {
-            return Response.ok(userService.getAllUsers()).build();
-        } catch (Exception e) {
-            return Response.status(Response.Status.INTERNAL_SERVER_ERROR).build();
+        boolean hasJWT = jwt.getClaimNames() != null;
+
+        if (hasJWT) {
+            try {
+                return Response.ok(userService.getAllUsers()).build();
+            } catch (Exception e) {
+                return Response.status(Response.Status.INTERNAL_SERVER_ERROR).build();
+            }
         }
+        return Response.status(Response.Status.UNAUTHORIZED).build();
     }
 
     @GET
-    @Path("/get/{userId}")
+    @Path("/get")
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
     @Authenticated
-    public Response getUserById(String userId) {
-        try {
-            return Response.ok(userService.getUserById(userId)).build();
-        } catch (UserNotFoundException e) {
-            return Response.status(Response.Status.NOT_FOUND).build();
+    public Response getUserById() {
+        boolean hasJWT = jwt.getClaimNames() != null;
+
+        if (hasJWT && jwt.containsClaim("memberId")) {
+            String userId = jwt.getClaim("memberId");
+            System.out.println(userId);
+            try {
+                return Response.ok(userService.getUserById(userId)).build();
+            } catch (UserNotFoundException e) {
+                return Response.status(Response.Status.NOT_FOUND).build();
+            }
         }
+        return Response.status(Response.Status.UNAUTHORIZED).build();
     }
+
 
     @POST
     @Path("/update")
@@ -80,14 +104,18 @@ public class UserController {
     public Response updateUser(@Context SecurityContext ctx, UpdateUserDTO userDTO) {
         Principal caller = ctx.getUserPrincipal();
         String name = caller == null ? "anonymous" : caller.getName();
+        boolean hasJWT = jwt.getClaimNames() != null;
 
-        try {
-            userService.updateUser(userDTO, name);
-            return Response.ok().build();
-        } catch (UserNotFoundException e) {
-            return Response.status(Response.Status.NOT_FOUND).build();
-        } catch (Exception e) {
-            return Response.status(Response.Status.INTERNAL_SERVER_ERROR).build();
+        if (hasJWT) {
+            try {
+                userService.updateUser(userDTO, name);
+                return Response.ok().build();
+            } catch (UserNotFoundException e) {
+                return Response.status(Response.Status.NOT_FOUND).build();
+            } catch (Exception e) {
+                return Response.status(Response.Status.INTERNAL_SERVER_ERROR).build();
+            }
         }
+        return Response.status(Response.Status.UNAUTHORIZED).build();
     }
 }
